requirements.txt
h5py==2.9.0
numpy==1.16.2
opencv-python==4.1.0.25
tensorboardX==1.6
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
Pillow==6.0.0